+++ CHANGELOG.md
@@ -12,6 +12,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
      * NewSessionTokenSignatureProviderFromFile
 - Cloud only: added support to read region from system environment variable
   OCI_REGION if using user principal or session token authentication
+- Latest Oracle Cloud Infrastructure regions and region codes: OZZ, DRS
+- Updated realm domain for OC22/NAP
 
 ## 1.4.1 - 2023-08-21